refactor: rename transaction toolbar and top band folders
