+++ frontend/src/transactions/components/Toolbar/hooks/useMobileSearchStuck.ts
